requirements.txt
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
flask-talisman==1.1.0
gunicorn==21.2.0
Werkzeug==3.0.1 